finish modify executor
